API/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pandas
python-dotenv